fix: Error Correction and note
Fix error in main that would return empty array and
txt file, now lets the user know it was not able to
get the recipe from the website.

Left a note in get_data.py to see if I can find out
why it returns everything good except Instructions
being none when the JSON is the same as other
websites. Only happens with some websites.